+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.4.0] 2025-11-05
+
 ### Added
  - `ctags` like handling for `-f <tagfile>` command line args param
  - Add support for reading command arguments from single file or directory with `.ctags` files